feat(config): add defaultPreviewSettings to seed initial preview-pane state
Histoire's preview-settings store is seeded with hardcoded defaults
(notably `responsiveWidth: 720`) the first time it runs. There is
currently no way to change those seeds from `histoire.config.js`, so
every story loads with the responsive picker constrained to 720px
even when the user wants previews to render at their natural width.

This commit exposes a new top-level config field
`defaultPreviewSettings: Partial<PreviewSettings>` that merges into the
useStorage initial value. Each field is optional; omitted fields fall
back to the built-in default. The user-written file `histoire.config.js`
becomes:

  export default defineConfig({
    defaultPreviewSettings: {
      responsiveWidth: null, // null → "Auto"
    },
  })

Only seeds the store on first visit. Once the user changes a setting
via the toolbar, their choice is persisted and wins on later loads.

Type changes:
- `PreviewSettings` moves from `packages/histoire-app/src/app/types.ts`
  to `packages/histoire-shared/src/types/config.ts` so the node-side
  config schema can reference it. The app-side type file re-exports.
- `responsiveWidth` and `responsiveHeight` widen from `number` to
  `number | null` so the picker UI's existing `?? 'Auto'` fallback
  has a typed home (the runtime already passed null for height; the
  number-only type was a latent mismatch).

Docs: new entry under `docs/reference/config.md` after
`defaultStoryProps`.
