+++ packages/histoire-shared/src/types/config.ts
@@ -28,6 +28,17 @@ export interface BackgroundPreset {
   contrastColor?: string
 }
 
+export interface PreviewSettings {
+  /** Width of the responsive preview. `null` renders as "Auto" (no constraint). */
+  responsiveWidth: number | null
+  /** Height of the responsive preview. `null` renders as "Auto" (no constraint). */
+  responsiveHeight: number | null
+  rotate: boolean
+  backgroundColor: string
+  checkerboard: boolean
+  textDirection: 'ltr' | 'rtl'
+}
+
 export interface TreeGroupConfig {
   title: string
   id?: string
@@ -181,6 +192,19 @@ export interface HistoireConfig {
    * @deprecated use `theme.darkClass` instead
    */
   sandboxDarkClass?: string
+  /**
+   * Initial values for the persisted preview-settings store. Lets users
+   * change the seed for the responsive size picker (which is otherwise
+   * hardcoded to 720px) or any other preview-pane UI setting. Each field
+   * is optional; omitted fields fall back to the built-in default. Set
+   * `responsiveWidth: null` to default the picker to "Auto" so previews
+   * render at their natural width.
+   *
+   * Only seeds the store on first visit (or when localStorage hasn't yet
+   * been written). Once the user changes a setting via the toolbar,
+   * their choice is persisted and takes precedence.
+   */
+  defaultPreviewSettings?: Partial<PreviewSettings>
   /**
    * Default props for stories.
    */
